requirements.txt
streamlit==1.23.0
pandas==1.5.3
numpy==1.24.3
plotly==5.14.1
requests==2.29.0
scipy==1.10.1
statsmodels==0.13.5
tensorflow==2.12.0
prophet==1.1.2
PyPDF2==3.0.1
python-dotenv==1.0.0
langchain==0.0.173
openai==0.27.8